+++ hs-nix-c-api/test/Test/Nix/C/Settings.hs
@@ -1,17 +1,17 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Test.Nix.Settings (spec) where
+module Test.Nix.C.Settings (spec) where
 
 import qualified Data.ByteString as BS
 import Test.Hspec
 
-import Nix.Context (NixError (..))
-import Nix.Unsafe.Init (initNix)
-import Nix.Unsafe.Settings (getSetting, setSetting)
+import Nix.C.Context (NixError (..))
+import Nix.C.Unsafe.Init (initNix)
+import Nix.C.Unsafe.Settings (getSetting, setSetting)
 
 spec :: Spec
-spec = describe "Nix.Settings" $ before_ initNix $ do
+spec = describe "Nix.C.Settings" $ before_ initNix $ do
   describe "getSetting" $ do
     it "reads a known setting" $ do
       val <- getSetting "store"